requirements.txt
python-dotenv
openai
PyMuPDF